+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRutils 1.7.2 - 2025-04-17
+* refactor `get_gDR_session_info` to display used packages 
+
 ## gDRutils 1.7.1 - 2025-04-16
 * synchronize Bioconductor and GitHub versioning